updated deleteGoal to ensure data is protect by user

+++ backend/controllers/goalControllers.js
@@ -1,6 +1,7 @@
 const asyncHandler = require('express-async-handler');
 
 const Goal = require('../models/goalModels');
+const User = require('../models/userModels');
 
 const getGoals = asyncHandler(async(req, res) => {
     const goals = await Goal.find({
@@ -35,13 +36,26 @@ const deleteGoals = asyncHandler(async(req, res) => {
         throw Error('Goal not found')
     }
 
+    const user = await User.findById(req.user.id)
+
+    // Check for User //
+
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // Make sure logged in user is goal user //
+
+    if(goal.user.toString() !== User.id) {
+        res.status(401)
+        throw new Error('User not authortized') 
+    }
+
     await Goal.deleteOne(goal)
 
     res.json({ id: req.params })
 
-    res.json({
-        message:`Delete goal ${req.params.id}`
-    })
 });
